Z_Tests/FlaskAPI/requirements.txt
flask
openai

python-dotenv


#pip install -r requirements.txt
#pip install --no-cache-dir -r requirements_test.txt